Got Textract Working in C#
Sadly it doesn't handle PDF's from a bytestream, but we do have images
working. We'll need to setup a S3 bucket for handling PDF files.
