+++ Appserver/appsettings.json
@@ -6,5 +6,9 @@
       "Microsoft.Hosting.Lifetime": "Information"
     }
   },
-  "AllowedHosts": "*"
+  "AllowedHosts": "*",
+  "AWS": {
+    "Profile": "default",
+    "Region": "us-west-2"
+  }
 }
